# docs/guide/public/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Grantline API",
    "version": "0.1.0-dev",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "description": "Single-organization, cookie-authenticated API. HTTPS sessions use Secure, HttpOnly, SameSite=Lax cookies. No personal API tokens in v0.1."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "session": []
    }
  ],
  "paths": {
...
    "/report": {
      "get": {
        "summary": "Export immutable report",
        "operationId": "get_report",
        "tags": [
          "report"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "/docs/reference/report.schema.json"
                }
              }
            }
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/graph": {
      "get": {
        "summary": "Bounded relationship neighborhood",
        "operationId": "get_graph",
        "tags": [
          "graph"
        ],
        "description": "Minimum role: viewer.",
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Structured error: 400 input; 401 session; 403 role/MFA/CSRF; 404 missing; 409 revision; 422 access test; 429 rate limit; 503 unavailable.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "run",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "entity",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "depth",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assertion",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "grantline_session"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}